feat: code-split VideoVault large files to satisfy G-SIZE02 [T001945]
