+++ VideoVault/server/handlers/movie-types.ts
@@ -1,0 +1,20 @@
+export interface MovieJobPayload {
+  inputPath: string;
+  movieId?: string;
+  rootKey?: string;
+  autoOrganize?: boolean;
+  baseDir?: string;
+}
+
+export interface MovieMetadata {
+  title: string;
+  year?: number;
+  duration: number;
+  width: number;
+  height: number;
+  bitrate: number;
+  codec: string;
+  fps: number;
+  aspectRatio: string;
+  fileSize: number;
+}
